# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BuyZone API",
    "description": "API for BuyZone application",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9009"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "paths": {
...
    "/api/products": {
      "get": {
        "tags": ["Products"],
        "summary": "Get products",
        "description": "User will receive all the products",
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/carts": {
      "post": {
        "tags": ["Carts"],
        "summary": "Add product to cart",
        "description": "User will add a product to cart",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "productID": {
                    "type": "string"
                  },
                  "quantity": {
                    "type": "number",
                    "format": "float"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/orders": {
      "get": {
        "tags": ["Orders"],
        "summary": "Receive cart products",
        "description": "User will receive cart products",

        "security": [
          {
            "JWT": {}
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    }
  }
}